move constants; fix total column to tally OT
