--- src/js/constants/app_constants.js
@@ -1,2 +1,0 @@
-// Set to equal current fiscal year
-export var FISCAL_YEAR = '26';
