+++ src/js/constants/excel_constants.js
@@ -1,4 +1,4 @@
-import { FISCAL_YEAR } from "./app_constants";
+import { FISCAL_YEAR } from "./budget_constants";
 
 // sheets to expect on detail sheet
 export const SHEETS = {
@@ -8,27 +8,12 @@ export const SHEETS = {
     'Revenue' : 'revenue'
 }
 
-// object categories (from obj part of account string)
-export const OBJ_CATEGORIES = {
-    list : [
-        // 'Salaries & Wages',
-        // 'Employee Benefits',
-        'Professional & Contractual Services',
-        'Operating Supplies',
-        'Operating Services',
-        'Equipment Acquisition',
-        'Capital Outlays',
-        'Fixed Charges',
-        'Other Expenses'
-    ]
-}
-
 // where to find the general fund target
 export const TARGET_CELL_ADDRESS = 'C14'
 
 export const TOTAL_COLUMNS = {
     'personnel': 'Total Sal/Wag & Ben Request',
-    'overtime':'Total Cost (including benefits)',
+    'overtime':`FY${FISCAL_YEAR} Total OT/SP/Hol + FICA Request`,
     'nonpersonnel': `FY${FISCAL_YEAR} Departmental Request Total`,
     'revenue': `FY${FISCAL_YEAR} Departmental Estimate`
 };
