+++ src/js/constants/index.js
@@ -1,2 +1,2 @@
-export * from './app_constants';
+export * from './budget_constants';
 export * from './excel_constants';
